Geometries: move all other headers as well
